+++ approvaltests-util-tests/src/test/java/com/spun/util/io/InternetConnectivityTest.java
@@ -1,4 +1,4 @@
-package com.spun.util.io.tests;
+package com.spun.util.io;
 
 import static org.junit.Assert.assertNull;
