+++ approvaltests-util-tests/src/test/java/com/spun/util/io/XMLUtilsTest.java
@@ -1,4 +1,4 @@
-package com.spun.util.io.tests;
+package com.spun.util.io;
 
 import org.w3c.dom.Document;
