+++ approvaltests-util-tests/src/test/resources/com/spun/util/io/Book1.csv
